web/requirements.txt
Django==1.11.6
gunicorn==19.7.1
psycopg2==2.7.3.1
djangorestframework==3.7.0
djangorestframework-jwt==1.11.0
